requirements.txt
pyyaml>=6.0.1
jinja2>=3.0.0
psycopg2-binary>=2.9.0
ruff>=0.3.0 